Bytecode VM: Move BytecodeLoader into a subdirectory
We were getting too many Bytecode classes, try to organize them a bit.
