+++ lib/natalie/compiler.rb
@@ -2,7 +2,7 @@ require 'tempfile'
 require_relative '../../build/generated/numbers'
 require_relative './compiler/backends/cpp_backend'
 require_relative './compiler/bytecode'
-require_relative './compiler/bytecode_loader'
+require_relative './compiler/bytecode/loader'
 require_relative './compiler/bytecode_ro_data'
 require_relative './compiler/comptime_values'
 require_relative './compiler/instruction_manager'
